fix: 修复简报页闪烁死循环和 sleepHours 精度丢失
- share.tsx: useFocusEffect 改用 errorRef+silentRefreshedRef 避免依赖 error state
  导致的死循环闪烁；所有 setError 调用同步更新 errorRef
- drizzle/schema.ts: sleepHours 从 int 改为 float，支持 9.5 等小数睡眠时间

+++ drizzle/schema.ts
@@ -1,4 +1,4 @@
-import { int, mysqlEnum, mysqlTable, text, timestamp, varchar, boolean, json } from "drizzle-orm/mysql-core";
+import { int, float, mysqlEnum, mysqlTable, text, timestamp, varchar, boolean, json } from "drizzle-orm/mysql-core";
 
 // ─── Users (existing) ────────────────────────────────────────────────────────
 
@@ -85,7 +85,7 @@ export const checkIns = mysqlTable("check_ins", {
   authorUserId: int("authorUserId").notNull(),
   date: varchar("date", { length: 10 }).notNull(),         // YYYY-MM-DD
   // Morning check-in
-  sleepHours: int("sleepHours"),
+  sleepHours: float("sleepHours"),
   sleepQuality: mysqlEnum("sleepQuality", ["poor", "fair", "good"]),
   sleepInput: json("sleepInput"),
   sleepScore: int("sleepScore"),
